+++ rollup/sync_service/eth_client.go
@@ -1,4 +1,4 @@
-package l1_msg
+package sync_service
 
 import (
 	"context"
